fix(errors): stack trace dumping after interaction timeout
Dynamic debugging typically causes the interaction/webhook to time out,
which sometimes leads to a failure to upload a stack trace. This change
anticipates interaction expiry and sends a new message instead. In the
event that the three previous attempts fail, we will log an internal
error instead (since a stack trace at this point is not informative.)

Signed-off-by: little-meow-meow <204685920+little-meow-meow@users.noreply.github.com>